added animated water with no seams thats follows the camera

+++ src/rendering/Model.cpp
@@ -58,19 +58,19 @@ VulkanModel::VulkanModel (VulkanRenderer& renderer, std::shared_ptr<MeshData> me
 
 	vertexCount = (uint32_t)mesh->vertexData.size ();
 	vertexElementCount = (uint32_t)mesh->desc.ElementCount ();
-	indexCount = (uint16_t)mesh->indexData.size ();
+	indexCount = (uint32_t)mesh->indexData.size ();
 
 	uint32_t vBufferSize = static_cast<uint32_t> (vertexCount) * sizeof (float);
-	uint16_t iBufferSize = static_cast<uint16_t> (indexCount) * sizeof (uint16_t);
+	uint32_t iBufferSize = static_cast<uint32_t> (indexCount) * sizeof (uint32_t);
 
 
 	vmaVertices = std::make_unique<VulkanBufferVertex> (renderer.device, vertexCount);
-	vmaIndicies = std::make_unique<VulkanBufferIndex> (renderer.device, (uint16_t)indexCount);
+	vmaIndicies = std::make_unique<VulkanBufferIndex> (renderer.device, (uint32_t)indexCount);
 
 	auto vertexStagingBuffer = std::make_shared<VulkanBufferStagingVertex> (
 	    renderer.device, mesh->vertexData.size (), mesh->vertexData.data ());
 	auto indexStagingBuffer = std::make_shared<VulkanBufferStagingIndex> (
-	    renderer.device, (uint16_t)mesh->indexData.size (), mesh->indexData.data ());
+	    renderer.device, (uint32_t)mesh->indexData.size (), mesh->indexData.data ());
 
 	VkBuffer vert = vmaVertices->buffer.buffer;
 	VkBuffer index = vmaIndicies->buffer.buffer;
